Update tablewriter to 1.0 (#3)
* Make package work with updated version v1.0.0 of tablewriter.

tablewriter has changed a lot of formatting behavior. This commit tries to emulate the old behavior as much as possible. It may not be complete though.

* Move handling of all the pretty-table stuff into its own file.

This declutters html2text.go

* Add the possibility to bypass all the legacy tablewriter handling and do all the configuration directly.

The place for the configuration function is slightly odd, but avoids larger refactoring and possible breaking changes for consumers.

* Update deps

Signed-off-by: James Hillyerd <james@hillyerd.com>

---------

Signed-off-by: James Hillyerd <james@hillyerd.com>
Co-authored-by: René 'Necoro' Neumann <necoro@necoro.eu>